+++ nokogumbo-import/lib/nokogumbo.rb
@@ -1,7 +1,6 @@
 require 'nokogiri'
 require 'nokogumbo/version'
 require 'nokogumbo/html5'
-require 'nokogumbo/xml/node.rb'
 
 require 'nokogumbo/nokogumbo'